Rename width/height
Rename all variables and parameters to something more descriptive to
stress differences. The new names are display_width, arena_width and
texture_width (and height, respectively).

There is a logical difference between the number of cells in the battle
arena and the number of pixels used to display them (even though, as of
now, they are the same).