Address particles by positions
Blob now maintains a hash map of the particles keyed on their positions.
This made operator== and a hash function necessary in Vector (to use
them as keys).

The hash function uses the lower half of the both components, meaning
that, as long as coordinates stay small (i.e. use at most half the bits)
and positive, the hash should be unique.