Make Direction a class
Previously an enum, Direction is now a class. That allows nice things
like direction.opposite() or direction.vector(), the latter of which
returns a vector pointing in that directions.

The direction itself is represented as an unsigned integer, so it can
still be used as an array index (there is a conversion operator).

Everything in there is a constexpr, so there shouldn't be any runtime
overhead.

Had to make some of the Vector constructors constexpr as well.