Use target
The pressure is now determined by the target that's been set for a
particle. In each step, a pressure vector pointing to the target is
added to the pressure, so that particles move towards their target.

Particles will only move to another position, if that position is empty.
Otherwise, they collide with the particle their, currently passing on
their entire pressure to the other particle.

When a particle has moved, its pressure is reduced to zero.