Avoid bubbles
When a particle is moved, all particles "behind it" (in movement
direction) are now moved in the same direction. This prevents bubbles
from appearing inside the blob. However, particles at the edges can get
separated if their neighboring particles move in different directions.

No attention is paid to the pressure of the particles that are moved,
only that they're behind the one being moved.

At the moment, this looks a bit weird, there are some lines protruding
from the blob. I hope this will go away when the pressures are updated
properly according to the distance traveled.