Drag side particles after line
Moving lines of particles can leave particles isolated, like in this
example (X's being particles):

XXXXXXX
X

Moving the upper row to the right makes the lower particles lose
connection.

In an attempt to fix this, particles to the side of the last particle
being moved are now checked whether they have other neighbors, and if
not, are moved with the line. This doesn't always work though, because I
didn't consider cases like these:

   XXXXXXXX
XXXX

Moving the upper row to the right leaves a side particle that doesn't get moved
because it has another neighbor (which isn't connected to the rest of the blob
though).

Checking to how many particles the side particles has a path to is expensive and
complex.