Maintain connectivity information
Each particle is now aware for each of its neighbors whether it can
reach the entire blob via that neighbor, i.e. whether for there exists a
path (just following the neighbor pointers) to any particle in the blob
that goes via the neighbor and never needs to visit the original
particle again.

On construction, these flags are all set to true, which is correct as
long as the reachability graph is 2-connected (which it is for all but
edge case circular blobs that are currently possible). A construction
for general blob shapes is possible but more complicated.

Maintaining this information will help to avoid splitting the blob while
moving it.