Attempt to maintain connectivity
Moving particles can separate parts of the blob from the rest (i.e.
there is no path via the neighbor pointers anymore). This
(unsuccessfully) attempts to fix that by using the reachability
information added previously.

The reachability information was poorly thought out and doesn't work.
Consider this:

    XXXX
  XXX
  XXX

Moving the upper row to the right will separate the 2 blocks. But if the
top right particle is asked whether it's still connected to the rest of
the blob via other directions than to the north (where it lost
connection), it will say yes, because it still has a particle to the
west and south.

In order to use this, you'd have to do a breadth first search through
the blob (or sub-blob) to check whether the particles to the west and
south maybe need the top right particle for connectivity (which they
do). But that would require iterating potentially over the entire blob
(worst case).

What you really want is to start at the top right particle (the one of
which we want to know whether we have to drag it in behind the line we
moved) and trace the hull of the blob (all particles on the outside). If
you can reach the last particle of the top row, you still have
connection. You might have to trace both clockwise and counter-clockwise
for correctness, I'm not sure.

But I think I really should just focus on local connectivity, i.e. if
there isn't a path of at most a few hops, the particle should be dragged
in behind. That also makes more sense simulation-wise, as otherwise the
decision to drag or not could be up to whether there is a very long,
convoluted path that keeps the blob connected. Plus, you know,
performance.