Maintain connectivity
Threw out the dysfunctional previous approach of reachability
information, including the information in the particles and the
direction bitmasks.

With the new approach, after moving a line of particles, it's simply
checked if either of the particles to the sides of the last particle of
the line can still reach it locally, i.e. if it has a neighbor in the
direction in which the line moved (then there is a path to the last
particle of the line via that neighbor). If not, the particle is moved
to the position where the last particle of the line was previously. The
process is then repeated for neighbors to the sides of the particle that
was just moved.

I can't conclusively argue that this process always terminates, but it
seems to work out so far. My worry is that a cycle of particles could
appear that constantly drag each other. I think it's fine since the
preference of direction from which we take the particle is always
counter-clockwise (and so we're tracing the hull of the bit that has
become disconnected), but I can't prove it.

While connectivity now seems to be maintained completely, bubbles can be
seen to be forming. But these aren't bubbles appearing because a
particle moved and no other took its place. These occur because the blob
loops back on itself. They are fine, and indeed expected. The reason
they're appearing is because, for testing, each particle has the same
target pressure to the same target, so they're all trying to overtake
one another.