Fix liveloops
Unlike I previously assumed, infinite loops can occur when dragging
particles behind a line. Added a set that keeps track of all particles
that have already been moved so that no particle is considered for
moving twice. This means the funcion always terminates since there are
finitely many particles.

I believe the only situation where the liveloops could occur was when
there was a bubble (one that appeared from the blob folding back on
itself) the hull of which (the particles bordering on it) the function
traced as it moved particles behind. In that situation, it's ok to
simply stop when all neighbors are either locally connected or have been
moved before. Since the latter of these lie on the cycle around the
bubble, they have connectivity in the other direction.

Even if this would happen in other situations, it would still be fine to
not consider moving them, since they are either the very first particle
that has been moved (which has connectivity to the last particle of the
line that we're moving after), or they must have a path to that particle
through neighbors that have also already been moved.