Store pointer to blob state
Instead of having the blob state be a value member of the blob, we have
to store a shared pointer so we can actually inject it for mocking.

Set a default argument that simply makes a pointer to a new blob state,
so none of the clients need any updating.