Split particle functionality
Factored out the functionality concerning position as well as pressure
of particles into two separate classes. A particle has a member of each
and mostly delegates calls to them (though there is some overlap, e.g.
advance() requires access to the current position).

I'm not entirely sure this is the best way to go about things, but when
I tried to mock the previous Particle class, I found I had to provide
plenty of behavior concerning the position handling. With the new
design, I should be able to just add a member of the position attribute
class.