Store neighbors as weak_ptrs
Storing particle neighbors as shared_ptrs introduces circular
dependencies and thus memory leaks. Storing them as weak_ptrs should
clear that up.

This also means that, to return a shared_ptr of a neighbor, it has to be
created first (by locking the weak_ptr), so for every call to
getNeighbor() a shared_ptr is created and the refcount increased. No
more returning const references to shared_ptrs. Not sure if that'll have
an impact performance-wise.