Add more Blob tests
These are all the useful tests I can think of right now. Also added a
method to the fixture that fails a test if the blob isn't connected
(based on the particle and particle_map members of the fixture) and a
little ASCII art picture illustrating the test data available in the
fixture.

One of the tests actually fails: If on a loop (e.g. a 3x3 square of
particles with the middle one cut out) you move one of the sides, the
entire loop is moved in behind by dragParticlesBehindLine() to maintain
connectivity, including the line that was initially moved. This means it
could happen that you move a particle, which is immediately moved back.
I'm not sure this absolutely has to be prevented though, but I'm also
not sure this can't lead to infinite loops.

One possible solution would be to add all particles of the initially
moved line to the done_set and pass it to dragParticlesBehindLine(), but
is potentially wasteful. Another is to record the position of the first
particle of the line as well as the direction of the line and rely on
the drag-behind algorithm moving in direction priorities in
counter-clockwise order (i.e. tracing a hull) to prevent it from
crossing the line.