Don't touch lines when dragging particles
Fixes the test failure described in the previous commit. I chose the
solution adding the line to the done_set, I don't think the overhead is
huge (in fact I think it might be lower, because I'm saving a special
check whether we're crossing the line we're not supposed to move).