Prioritize high pressure
When advancing particles, the one with the highest pressure is now
handled first. Particles are moved until none of them have pressure
left.

This required a number of changes: BlobState now advances the particles,
and all at once before any of them are moved. Blob::advanceParticle()
has been renamed to handleParticle(), "advance" means "do everything
possible within this class to increase the time step".

BlobState can return the particle with the highest pressure. At the
moment, that's just using a linear max search. It's probably possible to
keep the particles in a max heap instead to speed this up. The problem
is that their pressure vectors may be changed without BlobState knowing
about it (since the particle vector must be accessible from the
outside), and so maintaining the heap is difficult.

Blob queries the highest pressure particle and terminates the step once
that particle's pressure is zero. I.e. it no longer asks the particle
whether it wants to move but decides for the particle.

This has also brought a problem: It doesn't work anymore. Infinite loops
occur, where the highest pressure particle pulls a line to one side,
then the particle on the other end of the line becomes the highest
pressure particle and pulls the same line in the opposite direction.

One possible solution would be to only use each particle once per time
step as the highest pressure particle. This would require either keeping
a record of all the used ones, or a full sort on the particle vector
before beginning the time step (also making the heap optimization
impossible). This could also become difficult once time steps are
parameterized on the actual time that has passed.

Another possible solution might be to use the pressure of the leading
particle as "fuel" for the entire line that it pulls behind it. That
way, the particle on the other end wouldn't have pressure in the
opposite direction added to it. This would require looking behind the
particle that wants to move and checking if it has enough pressure to
drag everything behind. This becomes especially complicated when you
factor in dragging particles behind a line that has moved. Basically,
you'd have to simulate the movement and only commit to it once you know
you have enough pressure. I think it would also make blobs less
"responsive". Movement would occur later than it does right now. I'm not
sure that it wouldn't destroy some of the dynamics of blobs (e.g. other
particles could rush into the space left by particles that were forcibly
moved and prevent them from moving back, thus pushing the blob forwards
more quickly, at least in a small region).