Make particles wait for followers
When particles get disconnected from the rest of the blob, they now wait
until the bubble behind them (where they came from) has been filled by
another particle. During that time, they continue to share pressure with
their followers. Once the bubble has been filled, all followers are
removed. This solves the problem of particles always being followed by
someone.

A current problem is that Particle::canMove() now also incorporates the
check for the bubble. First of all, that makes it non-const, which is a
bit ugly. More importantly, the time step now stop way too early,
because the particle with the highest pressure can't move because
there's a bubble. But there are lower-pressure particles that can, but
they're never considered because there's no distinction between
inability to move because of low pressure and because of a bubble.

Also started maintaining followers and leaders as vectors of plain
pointers, for when I get rid off all those shared pointers.