Compare mobility instead of pressure
When selecting the next particle to consider for movement, Blob now asks
for the particle with the highest mobility. It's esenntially the same as
the one with the highest pressure, except that particles that are
blocked (i.e. whose canMove() returns false) are treated as having no
pressure. This solves the problem of the time step quitting early
because the only particle it'll look at is blocked.

Also tried to clean up Particle::canMove() (not being const), but it's
not really possible with the current design. I have to tell the pressure
state at some point whether there is a neighbor in the direction a
bubble was last recorded.