Manage position and pressure in particle
The separation of position and pressure state doesn't make sense
anymore. I'd introduced it to simplify testing, but I'm just going to
have to deal with it. Particle now handles everything itself.

I apparently made some mistakes in transferring the functionality and
the bahavior is different now. But that's ok, I was going to change it
anyway.