Reevaluate follwership on advance
When advancing a particle, it now goes through all its leaders and
checks whether it should still follow it.

If their pressure vectors are pointing in different directions (i.e.
their inner product is negative), or if the leader is close to the
particle (currently if it's a direct neighbor), the followership is
dropped entirely.

Otherwise, if the vector towards the leader is pointing in a different
direction than the sum of both particle's pressure vectors, followership
is reversed. This is to make sure the particle that's ahead in movement
direction is the one being slowed down for the other's benefit.