Apply pressure towards leader
When a leader shares pressure with its followers, it will now apply it
in the direction towards itself rather than towards its target.

The overall result looks fairly ok, if a bit chaotic. There are some
particles that seem to split off very quickly when the target is
reached. I think these might be pushed back by a sort of accumulating
rebound effect (pressure that's passed on). This should be counteracted
by passing on part of the existing pressure to the followers, but that's
already happening. Maybe I need to share a bigger part of it, or maybe I
need more followers (not just the immediate neighbors, that could be
just one).