Add target pressure per time
The amount of pressure towards a particle's target that is added on
every advance() now depends on the time step. For this purpose, all of
the advance() functions take the amount of time that the simulation
should be advanced by. This makes it possible to call advance() at
varying intervals and still get the same results (at least as long as
the target pressure doesn't depend on the state of the game).