Add configuration constants
Added a Config class containing constants governing the behavior of a
blob. They are: how much pressure a leading particle passes on to
followers when they're first added, how much on every time step advance,
how much pressure a particle passes on when it collides with another,
and how much pressure is needed for a particle to be considered able to
move.

Especially the fraction of pressure passed on in a collision has a very
noticable impact. As it gets smaller, the blob is less jittery. The
previous default of 100% (i.e. all of it) was chaos, the new one of 50%
looks a lot smoother.