Test and fix Particle more
Adds the rest of the tests I could think of.

Also includes some fixes for discovered issues: When adding followers,
the leader wasn't recorded.

There was a problem when a particle's target was its own position
(division by zero since the norm would be zero).

Also addressed a similar issue that could occur if
Config::target_pressure_share should be set to zero (not that that makes
much sense, but it's legal).