Handle input
Battle now calls InputParser in each main loop iteration to get a list
of InputActions (so far ExitAction and MouseDownAction). MouseDownAction
only prints the click coordinates at the moment, but ExitAction (when
pressing Escape) exits.

Having the events processed only after rendering means there is a little
delay, although it's probably not going to be noticable. The InputParser
must be called from the main thread (SDL says so), so the rendering
would have to be stuffed in a separate thread, if ever.