Set target via mouse click
When clicking into the window, every blob (currently there's only the
one) gets its target set to that position (with a hardcoded pressure per
second of 20).

Screen does a conversion of display coordinates to logical (particle)
coordinates, which uses the current window dimensions, so it's possible
to resize the window without breaking this.

Setting a target in State and Blob is too imprecise. There might need to
be a way to regulate how much pressure per second is applied, and there
should be a way to limit the "budget" of pressure that can be applied.
Eventually, it would also be usefuly to not have to send the entire
blob, but to be able to split it up.