Fix BlobStateTest failures
In order to successfully call move() on a particle, that particle must
be able to move according to canMove(). Since I have to use a real
particle in the mock (because I don't want to reimplement everything),
that real particle has to think it can move. That means we have to set a
target and advance it before we ask it to move. There was a call-through
missing in the mock that's necessary for this to work. Honestly don't
know how this worked before.