Only set target for some particles
When calling setTarget() on Blob, you now have to select for which
particles the target gets set. This is done by specifying a center
around which particles within a radius are selected. Right now this is
hardcoded for testing.

There is a new overload of getParticles() in BlobState that selects
particles based on those criteria.

If the particle map in BlobState were two nested one-dimensional maps
(i.e. x- and y-coordinates) rather than a map on one two-dimensional
key, this could be more efficient.