Don't move particles out of bounds
Blob no longer attempts to move particles out of the arena bounds.
Instead, when a particle would go out of bounds following its desired
movement direction, its pressure in that direction is set to 0. This is
necessary in order for the main loop to advance (it proceeds until none
of the particles have pressure left).