Make State initiate particle movement
The loop taking care of particle movement during a time step is now
handled in State rather than Blob. This allows to support more than one
Blob at the same time.

State repeatedly gets the particle with the highest mobility out of all
the blobs until that particle can't move anymore. It checks if that
particle would collide with an arena boundary if it were allowed to
move, and can also check if it would collide with another blob. If so,
it tells its blob that the particle has collided with a wall. Otherwise,
it passes it off to the blob to handle the movement of the particle
internally.

The responsibility of blob remains to collide particles of the same blob
in order to pass along pressure, or to simply move the particle.
Advancing particles (i.e. increasing their pressure) as well as checking
collision with the arena is handled in State. This means Blob needs no
longer be aware of the arena dimensions.

When State acts on a particle, it calls a method of Blob. At the moment,
these are just call-throughs and State could call the particle directly,
but this could be useful in case Blob needs some control over the
modifications of its particles in the future.