Collide hostile particles
Particles of different blobs can no longer share a position. A collision
check is made by State before moving a particle. If there is a particle
of another blob in the direction the particle would move, it is instead
handled as though it hits a wall (i.e. all pressure in that direction is
killed).

Since only the neighboring coordinate is checked, this relies on Blob to
not move a particle more than one coordinate (i.e. add a unit vector).