Remove colliding particles
When hostile particles collide, they are now removed (as they are
"damaged" or "eaten up"). Since both particles of the two involved blobs
are removed at the same time, it doesn't matter how they move against
each other, and no one can win.

Once both blobs (or really just one of them) are empty, an assertion
failure occurs because of the empty sets. This will need to be
addressed.