Damage particles
Instead of immediately removing particles upon hostile collision, they
are now damaged and only removed when their health reaches 0.

Particles have a health which is decreased during collisions with a
particle from another blob. Each particle also has a strength, which
depends on how many friendly particles are in the area. The amount of
health reduction starts with a fixed number that varies with the
relative strengths of the involved particles.

The particle strength is simply the number of friendly particles within
the 5x5 square centered on the particle. The difference in particle
strengths is subtracted from the health reduction, so that the stronger
particle's health is reduced less.

The way health reduction is calculated probably needs refinement to make
it more fair. There are also a lot of hardcoded parameters.