Avoid double compilation
The static library Game containing most of the functionality was used to
share the functionality between the main application and the unit tests.
But everything still needed to be compiled twice, once for the Game
target, and once for the WoTMin2D target.

By making Game an OBJECT library (rather than STATIC), this is finally
no longer necessary. As I understand it, CMake doesn't really view it as
a packaged library, but as a collection of compiled files that can be
shared and linked in anywhere.